DCS-Nuance-Mix-Accelerator-main/hospitality-bot/requirements.txt
requests==2.28.2
pyodbc==4.0.35
openai==0.27.2
Flask[async]>=2.0.0
gevent
azure-communication-email
azure-communication-sms